继续调整

+++ ubuntu/24.04/darling/test.sh
@@ -4,14 +4,12 @@ set -e
 IMAGE_NAME="${1:-darling:noble-20260608}"
 DOCKER_CMD="${DOCKER_CMD:-docker}"
 
-# if [[ "${GITHUB_ACTIONS:-}" == "true" ]]; then
-#   $DOCKER_CMD run --rm \
-#     "${IMAGE_NAME}" \
-#     bash -lc 'set -e; command -v darling; dpkg-query -W | grep -i "^darling"; echo hello-from-darling-image'
-#   exit 0
-# fi
-
-
+if [[ "${GITHUB_ACTIONS:-}" == "true" ]]; then
+  $DOCKER_CMD run --rm \
+    "${IMAGE_NAME}" \
+    bash -lc 'set -e; command -v darling; dpkg-query -W | grep -i "^darling"; echo hello-from-darling-image'
+  exit 0
+fi
 
 $DOCKER_CMD  run --rm -it \
   --privileged \
@@ -21,7 +19,7 @@ $DOCKER_CMD  run --rm -it \
   --tmpfs /root:rw,exec,mode=755 \
   --tmpfs /tmp:rw,exec,mode=1777 \
   -v "$PWD":/work \
-  ghcr.io/zarraxx/darling:noble-20260608 \
+  ${IMAGE_NAME} \
   darling shell /bin/sh -lc 'uname -a; echo hello-from-darling'
